fix: move out RefreshTokenConfig
